src/test/java/com/blogspot/ostas/leetcode/all/medium/sort_characters_by_frequency/SolutionTest.java
package com.blogspot.ostas.leetcode.all.medium.sort_characters_by_frequency;

import org.junit.jupiter.api.Test;

import static org.assertj.core.api.Assertions.assertThat;

class SolutionTest {

    /*
      Example 1:
    Input: s = "tree"
    Output: "eert"
    Explanation: 'e' appears twice while 'r' and 't' both appear once.
    So 'e' must appear before both 'r' and 't'. Therefore "eetr" is also a valid answer.
    Example 2:
    Input: s = "cccaaa"
    Output: "aaaccc"
    Explanation: Both 'c' and 'a' appear three times, so both "cccaaa" and "aaaccc" are valid answers.
    Note that "cacaca" is incorrect, as the same characters must be together.
    Example 3:
    Input: s = "Aabb"
    Output: "bbAa"
    Explanation: "bbaA" is also a valid answer, but "Aabb" is incorrect.
    Note that 'A' and 'a' are treated as two different characters.
      Constraints:
    1 <= s.length <= 5 * 105
    s consists of uppercase and lowercase English letters and digits.
    */
    @Test
    void example_0() {
        var solution = new Solution();
        String s = "tree";
        String expected = "eert";
        var result = solution.frequencySort(s);
        assertThat(result).isEqualTo(expected);
    }

    @Test
    void example_1() {
        var solution = new Solution();
        String s = "cccaaa";
        String expected = "aaaccc";
        var result = solution.frequencySort(s);
        assertThat(result).isEqualTo(expected);
    }

    @Test
    void example_2() {
        var solution = new Solution();
        String s = "Aabb";
        String expected = "bbAa";
        var result = solution.frequencySort(s);
        assertThat(result).isEqualTo(expected);
    }

}
